Support Structs and Absolute Value Calculations (#3)
Previously, only native types were supported. This was limiting in that you couldn't reference properties of poses, translations, transformations, etc. Now all structs whose schema is captured in the wpilog file are supported in the same fashion as they are in AdvantageScope (the code is based on AdvantageScope's code).

Also adds support for calculations involving absolute values. This is useful when the magnitude of the value matters. For example, an offset where the minimum should be the value closest to zero and not the most negative.

Fix #1